requirements.txt
django-customvueadmin
